+++ pl_bolts/datamodules/__init__.py
@@ -23,7 +23,7 @@ try:
     from pl_bolts.datamodules.stl10_datamodule import STL10DataModule
     from pl_bolts.datamodules.vocdetection_datamodule import VOCDetectionDataModule
 
-    from pl_bolts.datamodules.kitti_dataset import KittiDataset
+    from pl_bolts.datasets.kitti_dataset import KittiDataset
     from pl_bolts.datamodules.kitti_datamodule import KittiDataModule
 except ModuleNotFoundError:
     pass
