+++ pl_bolts/datamodules/cifar10_datamodule.py
@@ -6,7 +6,7 @@ import torch
 from pytorch_lightning import LightningDataModule
 from torch.utils.data import DataLoader, random_split
 
-from pl_bolts.datamodules.cifar10_dataset import TrialCIFAR10
+from pl_bolts.datasets.cifar10_dataset import TrialCIFAR10
 from pl_bolts.transforms.dataset_normalizations import cifar10_normalization
 
 try:
