+++ pl_bolts/datamodules/imagenet_datamodule.py
@@ -8,7 +8,7 @@ from pl_bolts.transforms.dataset_normalizations import imagenet_normalization
 
 try:
     from torchvision import transforms as transform_lib
-    from pl_bolts.datamodules.imagenet_dataset import UnlabeledImagenet
+    from pl_bolts.datasets.imagenet_dataset import UnlabeledImagenet
 except ModuleNotFoundError:
     warn('You want to use `torchvision` which is not installed yet,'  # pragma: no-cover
          ' install it with `pip install torchvision`.')
