+++ pl_bolts/datamodules/ssl_imagenet_datamodule.py
@@ -4,7 +4,7 @@ from warnings import warn
 from pytorch_lightning import LightningDataModule
 from torch.utils.data import DataLoader
 
-from pl_bolts.datamodules.imagenet_dataset import UnlabeledImagenet
+from pl_bolts.datasets.imagenet_dataset import UnlabeledImagenet
 from pl_bolts.transforms.dataset_normalizations import imagenet_normalization
 
 try:
