+++ pl_bolts/datasets/base_dataset.py
