+++ pl_bolts/datasets/mnist_dataset.py
